Extract method input parsers
